echobrief-backend/pyproject.toml
[project]
name = "echobrief-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "authlib>=1.3.0",
    "bcrypt>=5.0.0",
    "celery>=5.6.0",
    "edge-tts>=7.2.7",
    "fastapi[standard]>=0.127.0",
    "feedparser>=6.0.12",
    "httpx>=0.25.0",
    "openai>=2.14.0",
    "passlib>=1.7.4",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-slugify>=8.0.4",
    "redis>=7.1.0",
    "sqlmodel>=0.0.29",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
